+++ resources/views/layouts/notus/partials/scripts.blade.php
@@ -1,8 +1,8 @@
 <!-- Vendor -->
-@vite('vendor/koneko/laravel-vuexy-website-layout-simple-koneko/resources/assets/templates/notus-tailwind/vendor/popper/popper.js')
+@vite('vendor/koneko/laravel-koneko-layout-small-templates/resources/assets/templates/notus-tailwind/vendor/popper/popper.js')
 @yield('vendor-script')
 
 <!-- Page Script -->
 @stack('page-script')
 
-@vite('vendor/koneko/laravel-vuexy-website-layout-simple-koneko/resources/assets/js/notus/app.js')
+@vite('vendor/koneko/laravel-koneko-layout-small-templates/resources/assets/js/notus/app.js')