+++ resources/views/layouts/notus/partials/styles.blade.php
@@ -1,8 +1,8 @@
 <!-- Vendor Styles -->
-@vite('vendor/koneko/laravel-vuexy-website-layout-simple-koneko/resources/assets/notus/vendor/fonts/fontawesome.scss')
+@vite('vendor/koneko/laravel-koneko-layout-small-templates/resources/assets/notus/vendor/fonts/fontawesome.scss')
 @yield('vendor-style')
 
 <!-- Page Styles -->
 @stack('page-style')
 
-@vite('vendor/koneko/laravel-vuexy-website-layout-simple-koneko/resources/assets/css/notus/app.css')
+@vite('vendor/koneko/laravel-koneko-layout-small-templates/resources/assets/css/notus/app.css')