+++ resources/views/templates/samuel-coming-soon/page.blade.php
@@ -1,8 +1,8 @@
-@extends('vuexy-website-layout-simple-koneko::layouts.samuel-coming-soon.master')
+@extends('koneko-layout-small-templates::layouts.samuel-coming-soon.master')
 
 @section('content')
     <div class="relative h-screen w-full flex items-center justify-center bg-cover bg-center text-center px-5"
-        style="background-image:url({{ asset('vendor/vuexy-layout-simple-koneko/img/bg-coming-soon.png') }});">
+        style="background-image:url({{ asset('vendor/koneko-layout-small-templates/img/bg-coming-soon.png') }});">
         <div class="absolute top-0 right-0 bottom-0 left-0 bg-gray-900 opacity-75"></div>
 
         <div class="z-50 flex flex-col justify-center text-white w-full h-screen">